seperate Dark mode
add darkmode component to pages
