+++ src/App.js
@@ -3,14 +3,12 @@ import Countries from "./components/Countries";
 import SingleCountry from "./components/SingleCountry";
 
 
-
-
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Countries/>}></Route>
-        <Route path="/:name" element={<SingleCountry/>}></Route>
+        <Route path="/" element={<Countries />}></Route>
+        <Route path="/:name" element={<SingleCountry />}></Route>
       </Routes>
     </BrowserRouter>
   );
